*** empty log message ***
git-svn-id: http://code.jaspersoft.com/svn/repos/jasperreports/trunk@1347 76e25fc1-081a-0410-bb1b-c96bc70638b7

+++ jasperreports/src/net/sf/jasperreports/engine/fill/JRBigDecimalIncrementerFactory.java
@@ -258,7 +258,7 @@ class JRBigDecimalDistinctCountIncrementer extends JRAbstractExtendedIncrementer
 		DistinctCountHolder holder = 
 			(DistinctCountHolder)valueProvider.getValue(calculable.getHelperVariable(JRCalculable.HELPER_COUNT));
 		
-		return new Integer((int)holder.getCount());
+		return new BigDecimal(holder.getCount());
 	}
 	
 	public Object initialValue()
